Added edit/delete for audit records
